requirements.txt
# gRPC Python зависимости
grpcio>=1.75.0
grpcio-tools>=1.75.0

# Дополнительные зависимости для разработки
protobuf>=6.31.0
